requirements.txt
streamlit~=1.34.0
google-generativeai
streamlit
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
st-theme-1.2.3
streamlit-navigation-bar-3.3.0
